// dashboard/src/utils/hasura-api/generated/schemas/exportMetadataResponseMetadataSourcesItemTablesItemObjectRelationshipsItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { RelationshipUsing } from './relationshipUsing';

export type ExportMetadataResponseMetadataSourcesItemTablesItemObjectRelationshipsItem =
  {
    /** Name of the relationship */
    name?: string;
    using?: RelationshipUsing;
  };